chore: Update npm dependencies for Tailwind CSS and Flowbite
This commit adds the latest stable versions of Tailwind CSS and Flowbite as dependencies in the `package.json` file. The `tailwindcss` dependency is updated to version `3.4.9`, and the `flowbite` dependency is updated to version `2.5.1`. This update ensures that the project is using the most recent versions of these libraries.

The `tailwind.config.js` file is also created, which includes the configuration for Tailwind CSS. It extends the theme's colors by adding custom color values for 'primary', 'secondary', and 'ash'.

Lastly, the `base.html` template is updated to include the necessary CSS and JavaScript files for Tailwind CSS and Flowbite. The CSS files are compressed using the `compress` template tag, and the JavaScript files are compressed using the `compress` template tag as well. This ensures that the project's assets are optimized for performance.

This commit improves the project's styling capabilities by updating the dependencies and configuring Tailwind CSS and Flowbite.