refactor: Update email verification code field to use default value
The code changes include adding a new migration file `0002_alter_emailverification_email_code.py` in the `core/migrations` directory. This migration alters the `email_code` field of the `EmailVerification` model in the `core` app. The `email_code` field is updated to have a default value of `'e0d8ab'` and a maximum length of 6 characters.

Based on the code changes, the suggested commit message is "refactor: Update email verification code field to use default value".