refactor: Update user avatar in navbar to use image tag instead of SVG
The code changes involve replacing the SVG element with an image tag in the navbar.html template. The image source is set to the avatar URL of the logged-in user. This change improves the accessibility and visual representation of the user avatar in the navigation bar.

Based on the code changes, the suggested commit message is "refactor: Update user avatar in navbar to use image tag instead of SVG".