fix: Add default factort for email code
The replaced default value was evaluated every
time `makemigrations` was run, which caused
new migrations to be created every time.

I replaced it with a default factory function
that is evaluated when the model is created.

I deleted the migrations that were created
when the default value was a string. And changed
all references to those migrations.